fix: Some renaming and edits, details screen instead of modal
