+++ src/screens/TripDetails/Details/index.tsx
@@ -16,6 +16,7 @@ import usePollableResource from '../../../utils/use-pollable-resource';
 import Trip from '../components/Trip';
 import {TripDetailsTexts, useTranslation} from '../../../translations';
 import {Close} from '../../../assets/svg/icons/actions';
+import {ArrowLeft} from '../../../assets/svg/icons/navigation';
 
 export type DetailsRouteParams = {
   initialTripPatterns: TripPattern[];
@@ -86,7 +87,7 @@ const Details: React.FC<Props> = (props) => {
           accessible: true,
           accessibilityRole: 'button',
           accessibilityLabel: t(TripDetailsTexts.header.leftButton.a11yLabel),
-          icon: <ThemeIcon svg={Close} />,
+          icon: <ThemeIcon svg={ArrowLeft} />,
         }}
         title={t(TripDetailsTexts.header.title)}
         style={styles.header}
