+++ src/screens/TripDetails/Details/utils.ts
