+++ src/screens/TripDetails/LocationRow.tsx
