+++ src/screens/TripDetails/Map/MapLabel.tsx
