+++ src/screens/TripDetails/Map/utils.ts
