+++ src/screens/TripDetails/SituationRow.tsx
