+++ src/screens/TripDetails/components/Trip.tsx
