+++ src/screens/TripDetails/components/TripLegDecoration.tsx
