+++ src/screens/TripDetails/components/TripRow.tsx
