+++ src/screens/TripDetails/components/TripSection.tsx
