+++ src/screens/TripDetails/components/TripSummary.tsx
