+++ src/screens/TripDetails/components/WaitSection.tsx
