+++ src/screens/TripDetails/utils.ts
